requirements.txt
asgiref==3.9.1
Brotli==1.1.0
cffi==2.0.0
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.60.1
gunicorn==23.0.0
openpyxl==3.1.5
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
pycparser==2.23
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-barcode==0.16.1
sqlparse==0.5.3
tinycss2==1.5.1
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.4.0
